+++ client/src/App.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { ApolloProvider } from '@apollo/client'
 import apolloClient from './createApolloClient'
 
-import Products from './components/Products'
+import { Products } from './pages'
 import './styles/styles.css'
 
 const App = () => {
