+++ client/src/components/Cart.js
@@ -1,13 +1,34 @@
 import React from 'react'
+import { useQuery, useMutation } from '@apollo/client'
 import { Button } from 'reactstrap'
+import { GET_CART, REMOVE_FROM_CART } from '../gql'
 
 const Cart = props => {
+  const { loading, data } = useQuery(GET_CART, { fetchPolicy: 'network-only' })
+
+  const [removeFromCart] = useMutation(REMOVE_FROM_CART, {
+    update(cache, { data: { removeFromCart } }) {
+      cache.writeQuery({
+        query: GET_CART,
+        data: { cart: removeFromCart }
+      })
+    }
+  })
+
+  if (loading) {
+    return (
+      <div>
+        Loading cart...
+      </div>
+    )
+  }
+
   return (
     <div className="cart">
       <p>Cart</p>
       <ul>
-        {props.items && props.items.map((item, i) => (
-          <li key={i}>{item.name} <Button color="danger" onClick={() => props.removeItem(item.id)}>X</Button></li>
+        {!loading && data && data.cart.map((item, i) => (
+          <li key={i}>{item.name} <Button color="danger" onClick={() => removeFromCart({ variables: { productId: item.id }})}>X</Button></li>
         ))}
       </ul>
     </div>
