+++ client/src/components/index.js
@@ -1,0 +1,3 @@
+export { default as Product } from './Product'
+export { default as Sidebar } from './Sidebar'
+export { default as Cart } from './Cart'
