+++ client/src/gql/index.js
@@ -1,0 +1,2 @@
+export * from './cart'
+export * from './products'
